backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# LangChain ecosystem
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.15

# Ollama
ollama==0.1.6

# Document processing
pypdf==3.17.4
python-docx==1.1.0
python-multipart==0.0.6

# Vector database
pgvector==0.2.4

# Web search
tavily-python==0.3.3

# Token counting
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
